% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairyfloss.R
\docType{data}
\name{theme_fairyfloss}
\alias{theme_fairyfloss}
\alias{palette_fairyfloss}
\title{Fairyfloss theme}
\format{
An object of class \code{character} of length 9.
}
\usage{
theme_fairyfloss(
  base_size = 16,
  base_family = "Courier",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

palette_fairyfloss
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A very cute ggplot2 theme based off \href{https://twitter.com/sailorhg}{sailorhg}'s \href{https://sailorhg.github.io/fairyfloss/}{fairyfloss} text editor theme
}
\examples{
library(ggplot2)
ggplot(nintendo_sales, aes(x = sales_million, y = console)) +
  geom_col() +
  facet_wrap(~sales_type) +
  theme_fairyfloss()
}
\keyword{datasets}
