% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_empty.R
\name{gg_empty}
\alias{gg_empty}
\title{gg_empty}
\usage{
gg_empty(
  text = NULL,
  text_size = 12,
  title = NULL,
  title_size = 8,
  lib_path = NULL,
  safer_check = TRUE,
  error_text = ""
)
}
\arguments{
\item{text}{character string of the message to display.}

\item{text_size}{numeric value of the text size (in points).}

\item{title}{character string of the graph title.}

\item{title_size}{numeric value of the title size (in points).}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{c(lib_path, .libPaths())}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{safer_check}{Single logical value. Perform the "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Warning: must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
an empty plot.
}
\description{
display an empty ggplot2 plot with a text in the middle of the window (for instance to specify that no plot can be drawn).
}
\examples{
# simple example.
gg_empty(text = "NO GRAPH")
# white page.
gg_empty()
# all the arguments.
gg_empty(text = "NO GRAPH", text_size = 8, title = "GRAPH1", title_size = 10, lib_path = NULL)
}
