% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_palette.R
\name{gg_palette}
\alias{gg_palette}
\title{gg_palette}
\usage{
gg_palette(
  n,
  kind = "std",
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{n}{Single integer value indicating the number of colors required.}

\item{kind}{Single string. Either "std" for standard gg colors, "dark" for darkened gg colors, or "light" for pastel gg colors.}

\item{safer_check}{Single logical value. Perform some "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{unique(c(new, .Library.site, .Library))}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}. Of note, in \code{arg_check()}, \code{error_text} is also used at the end of the string returned when no problem is detected.}
}
\value{
The vector of hexadecimal colors.
}
\description{
provide colors used by ggplot2.
the interest is to use another single color that is not the red one used by default.
for ggplot2 specifications, see: https://ggplot2.tidyverse.org/articles/ggplot2-specs.html.
}
\examples{
gg_palette(n = 2) # output of the function
plot(1:7, pch = 16, cex = 5, col = gg_palette(n = 7)) # the ggplot2 palette when asking for 7 different colors
plot(1:7, pch = 16, cex = 5, col = gg_palette(n = 7)[5]) # selection of the 5th color of the ggplot2 palette made of 7 different colors
plot(1:7, pch = 16, cex = 5, col = gg_palette(n = 7, kind = "dark")) # the ggplot2 palette made of 7 darkened colors
plot(1:7, pch = 16, cex = 5, col = gg_palette(n = 7, kind = "light")) # the ggplot2 palette made of 7 lighten colors
}
