% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girlboss.R
\name{girlboss_col}
\alias{girlboss_col}
\title{Girlboss colours}
\usage{
girlboss_col(...)
}
\arguments{
\item{...}{}
}
\description{
Girlboss colours
}
\examples{
library(ggplot2)
n <- length(girlboss_col())
ggplot(data.frame(x = seq(1,n,1), y = rep(1,n), girlboss = girlboss_col()),
       aes(x,y,fill=girlboss)) +
  geom_tile() +
  scale_fill_manual(values = as.character(girlboss_col())) +
  theme_void() +
  coord_fixed(3)
}
