% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairyfloss.R
\name{scale_fill_fairyfloss}
\alias{scale_fill_fairyfloss}
\title{Fairyfloss fill scale}
\usage{
scale_fill_fairyfloss(discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Whether the colour aesthetic is discrete or not}

\item{reverse}{Whether the palette should be reversed}

\item{...}{Additional arguments}
}
\description{
Fairyfloss fill scale
}
\examples{
library(ggplot2)
ggplot(nintendo_sales, aes(x = sales_million, y = console, fill = sales_type)) +
  geom_col() +
  scale_fill_fairyfloss() +
  theme_fairyfloss()
}
