% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarpill.R
\name{scale_fill_sugarpill}
\alias{scale_fill_sugarpill}
\title{Sugarpill fill scale}
\usage{
scale_fill_sugarpill(discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Whether the colour aesthetic is discrete or not}

\item{reverse}{Whether the palette should be reversed}

\item{...}{Additional arguments}
}
\description{
Sugarpill fill scale
}
\examples{
library(ggplot2)
ggplot(nintendo_sales, aes(x = sales_million, y = console, fill = sales_type)) +
  geom_col() +
  scale_fill_sugarpill() +
  theme_sugarpill()
}
