% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girlboss.R
\name{theme_girlboss}
\alias{theme_girlboss}
\title{Girlboss theme}
\usage{
theme_girlboss(
  base_size = 16,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A very cute ggplot2 theme based off \href{https://twitter.com/sailorhg}{sailorhg}'s \href{https://sailorhg.github.io/fairyfloss/}{fairyfloss} text editor theme
}
\examples{
library(ggplot2)
ggplot(nintendo_sales, aes(x = sales_million, y = console)) +
  geom_col() +
  facet_wrap(~sales_type) +
  theme_girlboss()
}
