% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarpill.R
\name{theme_sugarpill}
\alias{theme_sugarpill}
\title{Sugarpill Fun Size theme}
\usage{
theme_sugarpill(
  base_size = 12,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A very cute ggplot2 theme based off the Sugarpill Fun Size Palette
}
\examples{
library(ggplot2)
ggplot(nintendo_sales, aes(x = sales_million, y = console, fill = console)) +
  geom_col() +
  facet_wrap(~sales_type) +
  scale_fill_sugarpill() +
  theme_sugarpill()
}
